refactor(domain): align terminology with ubiquitous language (glossary.md)
- Rename measurements.py to activity_values.py
  - Add backward-compat shim with deprecation warning
  - Update all imports across codebase

- Rename ErrorHandler to ErrorService
  - Add deprecated ErrorHandler alias with warning
  - Aligns with glossary rule: avoid "Handler", use "Service"

- Add deprecation warning to Work alias
  - Work -> PublicationEntity (CrossRef term -> domain term)

- Update terminology linter script
  - Add 'chunk' -> 'batch' rule
  - Update allowed files for activity_values.py

- Update architecture tests
  - Rename exemption entries for renamed files
  - Add activity.py to exemptions

All deprecated aliases emit DeprecationWarning for backward compatibility.
