+++ src/bioetl/application/services/__init__.py
@@ -30,6 +30,13 @@ from bioetl.application.services.medallion_lifecycle import (
     ClearResult,
     MedallionLifecycleService,
 )
+from bioetl.application.services.pipeline_runner_service import (
+    PipelineNotFoundError,
+    PipelineRunnerService,
+    RunOptions,
+    RunResult,
+    RunStatus,
+)
 from bioetl.application.services.quarantine_service import (
     PurgeResult,
     QuarantineRecord,
@@ -40,13 +47,6 @@ from bioetl.application.services.shutdown_service import (
     PipelineShutdownError,
     ShutdownReason,
     ShutdownService,
-)
-from bioetl.application.services.pipeline_runner_service import (
-    PipelineNotFoundError,
-    PipelineRunnerService,
-    RunOptions,
-    RunResult,
-    RunStatus,
 )
 from bioetl.application.services.vacuum_service import (
     TableCollectorPort,
