+++ src/bioetl/domain/__init__.py
@@ -40,29 +40,29 @@ from bioetl.domain.entities import (
     # DTO Records (Pydantic)
     ActivityRecord,
     ArticleRecord,
-    AssayRecord,
-    CellLineRecord,
-    DocumentRecord,
-    MoleculeRecord,
-    PubChemCompoundRecord,
-    PublicationRecord,
-    TargetComponentRecord,
-    TargetRecord,
     # Domain Entities (dataclass)
     Assay,
+    AssayRecord,
     BaseEntity,
     Bioactivity,
     BioactivityState,
     CellLine,
+    CellLineRecord,
     Compound,
     CompoundRecord,
     Document,
+    DocumentRecord,
     Molecule,
+    MoleculeRecord,
     Protein,
+    PubChemCompoundRecord,
     Publication,
     PublicationEntity,
+    PublicationRecord,
     Target,
     TargetComponent,
+    TargetComponentRecord,
+    TargetRecord,
     Work,  # Deprecated: use PublicationEntity
 )
