+++ src/bioetl/domain/ports/runner.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
 
 if TYPE_CHECKING:
-    from pathlib import Path
 
     from bioetl.domain.context import PipelineRunContext
