+++ tests/unit/application/services/test_pipeline_runner_service.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from datetime import UTC, datetime
 from unittest.mock import AsyncMock, MagicMock
-from uuid import UUID, uuid4
+from uuid import uuid4
 
 import pytest
